Factor out `randomDuration` from `controlplane` into `internal/timex` for re-use by other pollers
