+++ router/pkg/controlplane/poll_test.go
@@ -33,42 +33,3 @@ func Test_Poller(t *testing.T) {
 		assert.NoError(t, err)
 	})
 }
-
-func Test_RandomDuration(t *testing.T) {
-	t.Run("should return durations within acceptable range", func(t *testing.T) {
-		max := 10 * time.Millisecond
-
-		durations := sampleRandomDurations(25, max)
-
-		for _, duration := range durations {
-			assert.GreaterOrEqual(t, duration, 0*time.Millisecond)
-			assert.LessOrEqual(t, duration, max)
-		}
-	})
-
-	t.Run("should return 0 when max is 0", func(t *testing.T) {
-		max := 0 * time.Millisecond
-
-		duration := randomDuration(max)
-
-		assert.Equal(t, 0*time.Millisecond, duration)
-	})
-
-	t.Run("should panic when max is less than zero", func(t *testing.T) {
-		max := -1 * time.Millisecond
-
-		assert.Panics(t, func() {
-			randomDuration(max)
-		})
-	})
-}
-
-func sampleRandomDurations(count int, max time.Duration) []time.Duration {
-	durations := make([]time.Duration, count)
-
-	for i := 0; i < count; i++ {
-		durations[i] = randomDuration(max)
-	}
-
-	return durations
-}
